# [tabular] speed up tests