--- a/tabular/tests/unittests/calibrate/test_calibrate.py
+++ b/tabular/tests/unittests/calibrate/test_calibrate.py
@@ -4,21 +4,21 @@ def test_calibrate_binary(fit_helper):
         hyperparameters={"GBM": {}},
         calibrate=True,
     )
-    dataset_name = "adult"
+    dataset_name = "toy_binary"
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, sample_size=100)
+    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
 
 
 def test_calibrate_binary_bag(fit_helper):
     """Tests that calibrate=True doesn't crash in binary w/ bagging"""
     fit_args = dict(
-        hyperparameters={"GBM": {}},
+        hyperparameters={"GBM": {"ag_args_ensemble": {"fold_fitting_strategy": "sequential_local"}}},
         calibrate=True,
         num_bag_folds=3,
     )
-    dataset_name = "adult"
+    dataset_name = "toy_binary"
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, sample_size=100)
+    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
 
 
 def test_calibrate_multiclass(fit_helper):
@@ -27,21 +27,21 @@ def test_calibrate_multiclass(fit_helper):
         hyperparameters={"GBM": {}},
         calibrate=True,
     )
-    dataset_name = "covertype_small"
+    dataset_name = "toy_multiclass"
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, sample_size=100)
+    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
 
 
 def test_calibrate_multiclass_bag(fit_helper):
     """Tests that calibrate=True doesn't crash in multiclass w/ bagging"""
     fit_args = dict(
-        hyperparameters={"GBM": {}},
+        hyperparameters={"GBM": {"ag_args_ensemble": {"fold_fitting_strategy": "sequential_local"}}},
         calibrate=True,
         num_bag_folds=3,
     )
-    dataset_name = "covertype_small"
+    dataset_name = "toy_multiclass"
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, sample_size=100)
+    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
 
 
 def test_calibrate_quantile(fit_helper):
@@ -50,20 +50,18 @@ def test_calibrate_quantile(fit_helper):
         hyperparameters={"RF": {}},
         calibrate=True,
     )
-    dataset_name = "ames"
-    init_args = dict(problem_type="quantile", quantile_levels=[0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9])
+    dataset_name = "toy_quantile"
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, init_args=init_args, sample_size=100)
+    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)
 
 
 def test_calibrate_quantile_bag(fit_helper):
     """Tests that calibrate=True doesn't crash in quantile w/ bagging"""
     fit_args = dict(
-        hyperparameters={"RF": {}},
+        hyperparameters={"RF": {"ag_args_ensemble": {"fold_fitting_strategy": "sequential_local"}}},
         calibrate=True,
         num_bag_folds=3,
     )
-    dataset_name = "ames"
-    init_args = dict(problem_type="quantile", quantile_levels=[0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9])
+    dataset_name = "toy_quantile"
 
-    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args, init_args=init_args, sample_size=100)
+    fit_helper.fit_and_validate_dataset(dataset_name=dataset_name, fit_args=fit_args)