*Issue #, if available:*

Same as #4942, moved code to autogluon branch so unit tests work correctly in CI

*Description of changes:*

# Tabular Tests Speedup: 3060s -> 743s (4x speedup)

Massively speed up unit tests for tabular models


- Switch from real datasets to toy synthetic datasets
- Use toy model hyperparameters when faster
- Avoid calling `predictor.info()` for each model fit
- Avoid calling `predictor.leaderboard(..., extra_info=True)` for each model fit
- Only call refit_full for a given model family once, no need to test refit_full on every dataset
- Only test bagging for a given model family once, no need to test bagging on every dataset
- Ensure we test bagging for every model family (previously was inconsistent)
- Ensure we are testing every supported problem_type (previously it was inconsistent)

## Enhancements

- Add classmethod to each model to get the supported problem types

## Speedup

- AutoMM tests: 600s -> 60s
- LightGBM tests: 130s -> 16s 
- Dummy tests: 23s -> 12s
- NeuralNetTorch tests: 140s -> 8s
- NeuralNetFastAI tests: 15s -> 2s
- RandomForest tests: 87s -> 8s
- ExtraTrees tests: 108s -> 7s
- XGBoost tests: 32s -> 4s
- CatBoost tests: 61s -> 3s
- Linear tests: 23s -> 6s
- test_tabular: 233s -> 25s
- test_advanced_functionality: 67s -> 33s
- test_tabularHPO: 199s -> 106s
- test_tabular_parallel_local_bagging: 38s -> 35s
- test_tabular_sequential_local_bagging: 16s -> 8s
- test_calibrate.py: 26s -> 11s
- test_learning_curves.py: 70s -> 40s
- test_pseudolabeling: 68s -> 36s
- test_tabularHPObagstack: 204s -> 51s
- test_tabular_bagstack: 169s -> 6s
- test_tabular_bagstack_use_bag_holdout: 168s -> 6s
- test_sample_weight: 52s -> 15s

## Fixes

- test_tabular_feature_prune wasn't actually testing feature pruning, now it does
- Fixed several bugs with oof predictions on quantile problem type

TODO:

- [ ] Verify categorical, missing value, etc. handling for models (Follow-up, currently still have some tests using real datasets to verify this)

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.