@Innixma 2025-02-28
Note: I am aware that the code in `test_tabular.py` is still messy (but a lot less messy than it was...).

I plan to do a follow-up PR to clean this up. I avoided doing that in this PR to avoid it being increasingly hard to review.